requirements.txt
aiogram_i18n==1.3.4
Babel==2.14.0
aiosqlite==0.20.0
sqlalchemy==2.0.29
aiogram_calendar=0.5.0
